requirements.txt
azure-functions==1.17.0
beautifulsoup4==4.12.2
certifi==2023.7.22
charset-normalizer==3.2.0
idna==3.4
requests==2.31.0
ruff==0.0.286
soupsieve==2.4.1
urllib3==2.0.4
